+++ utils/GeneralFunctions.cpp
@@ -32,7 +32,7 @@ QString getUserHomeDirectory() {
 
 QString parseFromText(const QString& text) {
     QString parsedText;
-    for(int i{0}; i < text.length(); i++) {
+    for(int i{ 0 }; i < text.length(); i++) {
         if(text.at(i) == '\\' && i < text.length() - 1) {
             if(text.at(i+1) == 'n') {
                 parsedText.append('\n');
@@ -48,7 +48,7 @@ QString parseFromText(const QString& text) {
 
 QString parseToText(const QString& text) {
     QString parsedText;
-    for(int i{0}; i < text.length(); i++) {
+    for(int i{ 0 }; i < text.length(); i++) {
         if(text == '\n') {
             parsedText.append("\\n");
         }
@@ -67,14 +67,14 @@ namespace SerializationUtils {
     void writeQString(std::ofstream& out, const QString& str) {
         // conversion so i can get size in bytes, not characters
         QByteArray utf8Data = str.toUtf8();
-        int size = utf8Data.size();
-        out.write(reinterpret_cast<const char*>(&size), sizeof(int));
+        size_t size = utf8Data.size();
+        out.write(reinterpret_cast<const char*>(&size), sizeof(size_t));
         out.write(utf8Data.constData(), size);
     }
 
     void readQString(std::ifstream& in, QString& str) {
-        int size;
-        in.read(reinterpret_cast<char*>(&size), sizeof(int));
+        size_t size;
+        in.read(reinterpret_cast<char*>(&size), sizeof(size_t));
         QByteArray utf8Data(size, Qt::Uninitialized);  // Reserve space for UTF-8 data
         in.read(utf8Data.data(), size);
         str = QString::fromUtf8(utf8Data);
